Cleanup test projects.
